Translated some of the decoherence methods, refined the hurface hop rescaling to ensure the correct state is used in the second kick, added computations of the time-overlaps from the diabatic-to-adiabatic transformation for the LD and related methods, added examples and unittests for decoherence methods, added an example of running Tully1 model with several TSH and decoherence methods

+++ src/libra_py/dyn/control_params.py
@@ -386,7 +386,7 @@ class DynControlParams:
     #:  7 DISH, rev2023
     #:  8 diabatic IDA, experimental
     #:  9 simple decoherence, experimental
-    decoherence_algo: float = -1
+    decoherence_algo: int = -1
 
     #: SDM norm tolerance
     #: Corresponds to the `tol` parameter in the sdm function. It controls
